requirements.txt
pygraphviz
pymongo
redis
pyvex
angr
claripy
numpy
networkx
intervaltree
r2pipe
progressbar2
pystache
nltk
pyenchant
sklearn
scipy
matplotlib
pandas
seaborn
pydot
lief
dill
multiprocess
capstone
annoy
numba
tensorflow
gensim
tqdm
suffix-trees
IPython
timeout-decorator
tkimg
tk-tools
dgl
torch
psycopg2-binary
coloredlogs
joblib
posix_ipc
